+++ src/pipes.h
@@ -3,11 +3,11 @@
 
 typedef struct _pipes * Pipes;
 
-Pipes pipes_create(int size);
+Pipes pipes_create(size_t size);
 
 void pipes_append(Pipes p);
 
-int* pipes_index(Pipes p, int idx);
+int* pipes_index(Pipes p, size_t idx);
 
 int* pipes_last(Pipes p);
